# swagger.yaml
definitions:
...
  Product:
    properties:
      date_created:
        description: Date of a product creation
        format: date-time
        type: string
        x-go-name: DateCreated
      date_updated:
        description: Date of a product last modification
        format: date-time
        type: string
        x-go-name: DateUpdated
      description:
        description: Description of a product
        type: string
        x-go-name: Description
      price:
        description: |-
          Price of a product

          gte:0.00
        format: float
        type: number
        x-go-name: Price
      product_id:
        description: UUID of a product
        type: string
        x-go-name: ID
      stock:
        description: |-
          Stock of a product

          gte:0
        format: int64
        type: integer
        x-go-name: Stock
      title:
        description: Title of a product
        type: string
        x-go-name: Title
    required:
    - title
    - description
    - price
    - stock
    title: Product is an particular product.
    type: object
    x-go-package: github.com/rtbe/clean-rest-api/domain/entity
  StringArray:
    items:
      type: string
    title: StringArray represents a one-dimensional array of the PostgreSQL character
      types.
    type: array
    x-go-package: github.com/lib/pq
...
  User:
    properties:
      date_created:
        description: Date of a user creation
        format: date-time
        type: string
        x-go-name: DateCreated
      date_updated:
        description: Date of a user last modification
        format: date-time
        type: string
        x-go-name: DateUpdated
      email:
        description: Email of a user
        example: user@google.com
        type: string
        x-go-name: Email
      first_name:
        description: First name of a user
        type: string
        x-go-name: FirstName
      last_name:
        description: Last name of a user
        type: string
        x-go-name: LastName
      password:
        description: Password of a user
        items:
          format: uint8
          type: integer
        type: array
        x-go-name: Password
      roles:
        $ref: '#/definitions/StringArray'
      user_id:
        description: The UUID of a user
        type: string
        x-go-name: ID
      user_name:
        description: Username of a user
        type: string
        x-go-name: UserName
    required:
    - user_id
    - user_name
    - password
    - email
    title: User is a particular user.
    type: object
    x-go-package: github.com/rtbe/clean-rest-api/domain/entity
info: {}
paths:
...
  /products/:
    post:
      consumes:
      - application/json
      description: |-
        Creates a new product
        .
      operationId: createProduct
      produces:
      - application/json
      responses:
        "201":
          description: Product
          schema:
            $ref: '#/definitions/Product'
        "500":
          $ref: '#/responses/errorResponse'
      tags:
      - product
  /products/{id}:
    delete:
      description: |-
        Deletes a product by it\`s id
        and returns it\`s JSON representation.
      operationId: deleteProduct
      produces:
      - application/json
      responses:
        "204":
          $ref: '#/responses/emptyResponse'
        "500":
          $ref: '#/responses/errorResponse'
      tags:
      - product
    get:
      description: |-
        Gets a product by it\`s id
        and returns it\`s JSON representation.
      operationId: getProduct
      produces:
      - application/json
      responses:
        "200":
          description: Product
          schema:
            $ref: '#/definitions/Product'
        "500":
          $ref: '#/responses/errorResponse'
      tags:
      - product
    patch:
      consumes:
      - application/json
      description: |-
        Updates a product
        .
      operationId: updateProduct
      responses:
        "201":
          $ref: '#/responses/emptyResponse'
        "500":
          $ref: '#/responses/errorResponse'
      tags:
      - product
  /products/{lastSeenID}/{id}:
    get:
      description: |-
        This request uses two provided values to implement pagination: last seen id and limit.
        Results of a request sorted by creation date of selected users and sended back as JSON.
      operationId: listProducts
      produces:
      - application/json
      responses:
        "200":
          description: Product
          schema:
            items:
              $ref: '#/definitions/Product'
            type: array
        "500":
          $ref: '#/responses/errorResponse'
      summary: Gets paginated list of products.
      tags:
      - product
...
  /users/{id}:
    delete:
      description: |-
        Deletes a user
        .
      operationId: deleteUser
      produces:
      - application/json
      responses:
        "204":
          $ref: '#/responses/emptyResponse'
        "500":
          $ref: '#/responses/errorResponse'
      tags:
      - user
    patch:
      consumes:
      - application/json
      description: |-
        Updates a user
        .
      operationId: updateUser
      responses:
        "204":
          $ref: '#/responses/emptyResponse'
        "500":
          $ref: '#/responses/errorResponse'
      tags:
      - user
  /users/{lastSeenID}/{limit}:
    get:
      description: |-
        This request uses two provided values to implement pagination: last seen id and limit.
        Results of a request sorted by creation date of selected users and sended back as JSON.
      operationId: listUsers
      produces:
      - application/json
      responses:
        "200":
          description: User
          schema:
            items:
              $ref: '#/definitions/User'
            type: array
        "500":
          $ref: '#/responses/errorResponse'
      summary: Gets paginated list of users.
      tags:
      - user
  /users/id/{id}:
    get:
      consumes:
      - application/json
      description: |-
        Gets a user by his id
        and returns it\`s JSON representation.
      operationId: getUser
      produces:
      - application/json
      responses:
        "201":
          description: User
          schema:
            $ref: '#/definitions/User'
        "500":
          $ref: '#/responses/errorResponse'
      tags:
      - user
responses:
  errorResponse:
    description: Empty response
    schema:
      properties:
        error:
          example: There are some error
          type: string
          x-go-name: Error
        fields:
          type: string
          x-go-name: Fields
      type: object
swagger: "2.0"